+++ dmr/throttling/algorithms.py
@@ -194,6 +194,7 @@ class LeakyBucket(BaseThrottleAlgorithm):
     ) -> CachedRateLimit:
         """Check access; raise when the bucket is full."""
         cache_object, now = self._process_cache(throttle, cache_object)
+        # First, decrease the usage level for the ellapsed time:
         elapsed = now - cache_object['time']
         level = max(
             0,
@@ -201,8 +202,14 @@ class LeakyBucket(BaseThrottleAlgorithm):
             cache_object['history'][0] - elapsed * throttle.max_requests,
         )
         cache_object = CachedRateLimit(history=[level], time=now)
+<<<<<<< Updated upstream
 
         if cache_object['history'][0] + throttle.duration_in_seconds > (
+=======
+        print(cache_object, throttle.max_requests * throttle.duration_in_seconds)
+        # Do the check:
+        if cache_object['history'][0] >= (
+>>>>>>> Stashed changes
             throttle.max_requests * throttle.duration_in_seconds
         ):
             raise TooManyRequestsError(
@@ -214,7 +221,7 @@ class LeakyBucket(BaseThrottleAlgorithm):
                     cache_object['time'],
                 ),
             )
-        # One scaled unit of a request:
+        # Add one scaled unit for the performed request:
         cache_object['history'][0] += throttle.duration_in_seconds
         return cache_object